+++ Inc/DCMotorHardware.h
@@ -29,7 +29,7 @@ public:
 			const int32_t motor_left_timer_channel);
 	~DCMotorHardware();
 
-	uint32_t getTicks(uint32_t encoderId);
+	int16_t getTicks(uint32_t encoderId);
 
 	/**
 	 *
